+++ CHANGELOG.md
@@ -15,4 +15,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Initial creation
 
 [Unreleased]: https://github.com/xmidt-org/pluginfx/compare/v0.0.1..HEAD
-[v0.0.1]: https://github.com/xmidt-org/pluginfx/compare/0.0.0...v0.0.1
+[v0.0.1]: https://github.com/xmidt-org/pluginfx/compare/v0.0.0...v0.0.1